qr_scanner/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dateutil==2.8.2
pillow==10.1.0
pyzbar==0.1.9
